# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": " register new account",
                "description": " ",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['الإيميل مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error schema",
                "properties": {
                    "title": {
                        "description": "Error Message",
                        "type": "string",
                        "example": "  error meassage "
                    },
                    "desc": {
                        "description": "Error description",
                        "type": "string",
                        "example": "Something went wrong, please try again later."
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "title": "Register Request",
                "description": "RegisterRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name ",
                        "type": "string",
                        "example": "osama almamari"
                    },
                    "email": {
                        "title": "email",
                        "description": "email ",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password  ",
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password_confirmation  ",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User Data",
                "description": "Schema for user data",
                "properties": {
                    "id": {
                        "description": "The ID of the user",
                        "type": "string",
                        "example": "01h84ym49w7byskc4wrcmnyszc"
                    },
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "example": "اسامة"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    },
                    "token": {
                        "description": "tokrn user",
                        "type": "string",
                        "example": "01h84ym49w7byskc4wrcmnyszcddddddddddddddddd"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}